ProjectMgtService-GettingStarted: Updating the bussiness model of the project.
